requirements.txt
django-4.2.14
pip install djangorestframework